Added a healthcheck to Redis, renamed services and cleared an unused secret in compose.yml
In the guide, they used service_healthy condition on Redis, so I thougt that Redis has a built-in healthcheck. That, seemingly, isn't the case (which makes service_healthy practically equivalent with service_started), so I supplied my own proper healthcheck.

As for the names - "app" is reserved for future use (as the main guziohub.ovh site) so I swapped it for (a much more descriptive, btw) "nextcloud" and "nextcloudcron" felt too long. I wish I could rename "imaginary" to "ncimages", but I'm pretty sure that name is hard-coded, due to how its based on its image the AIO image family (where, of course, all coontainers should be configured for you automatically, so they'd have no reason to give an option to customise their names - or, I guess, I didn't look into it enough)