Changed Minecraft's ownership structure and documented some of my earlier discoveries.
Apparently, manually setting UID+GID to their out-of-container values in-container doesn't worek very well under Podman becasue (unilke Docker, that (with its root-ful containers) maps IDs 1:1) it maps in-container 0 to out-of-container „you” and everything else in-container is mapped (including things with the same ID as you out-of-container) to some arbitraty unused out-of-container IDs (that are your sub-IDs, that is you `usermod`ed them onto yourself - and while saying this, I thought that I should probably document this). Which you, the user, can't access directly (they ARE your subs, so you could (I think???) `chown` them back, but that requires you to stop the containers first, unless you want to break permissions at runtime, which is not a fun time) - and that's the very opposite of what I aimed to achive by mapping those IDs onto myslef. So back to the root user they go!

I've known of this problem for some time now, but couldn't be bothered to do anything about it (it's not like I was actually running an MC server there, anyway). But I just got a nice opportunity for taking care of this: I changed folder names in the previous commit and (since there was nothing important in it, anyway) nuked the minecraft folder completely. It was the easiest way: it'll just automatically re-create with a new name. So I can use this re-creation to also give it a new UID+GID.