Changed gitignore
I was messing around with running an MC server on this, then I forgot about it, then I wanted to get back to working on my website, only to remember that I made a huge mess with modpacks. So I updated the gitignore to not have it included.

Also, since I not only have „the DB”, but also several sub-DBs for things that don't work with Postgres - I renamed „Datebase” to „Databases” to reflect that.